+++ app/components/folio/console/application_component.rb
@@ -3,4 +3,5 @@
 class Folio::Console::ApplicationComponent < Folio::ApplicationComponent
   include Folio::Console::PreviewUrlFor
   include Folio::Console::UiHelper
+  include Folio::StringHelper
 end