+++ app/helpers/folio/console/cells_helper.rb
@@ -33,9 +33,20 @@ module Folio
         opts[:right] = capture(&block)
       end
 
-      opts[:form_errors_shown] = @form_errors_shown
+      opts[:form_errors_shown] = @form_errors_shown # todo seems to be unused
 
-      cell("folio/console/form/header", f, opts).show.html_safe
+      # todo html_safe?
+      # cell("folio/console/form/header", f, opts).show.html_safe
+
+      render(Folio::Console::Form::HeaderComponent.new(model: f,
+                                                      title: opts[:title],
+                                                      title_class_name: opts[:title_class_name],
+                                                      subtitle: opts[:subtitle],
+                                                      left: opts[:left],
+                                                      right: opts[:right],
+                                                      sti_badge: opts[:sti_badge],
+                                                      tabs: opts[:tabs],
+                                                      hide_fix_error_btn: opts.slice(:hide_fix_error_btn)))
     end
   end
 end