+++ CHANGELOG.md
@@ -33,6 +33,8 @@ All notable changes to this project will be documented in this file.
 - `Folio::Console::Ui::AjaxInputComponent` now expects API to return `{ name => new_value }` instead of `{ value: new_value }`
 - redid input autocomplete without jQuery UI
 - default console `show` action now redirects to `edit` - make sure you have your `show` actions explicitely defined on controllers and are not using `super`
+- Cells -> ViewComponents refactoring:
+  - Folio::Console::Form::HeaderCell -> Folio::Console::Form::HeaderComponent
 
 ## [6.5.1] - 2025-06-18